update for Chapel 1.31

+++ apps/pagerank/pagerank.cu
@@ -53,7 +53,7 @@ __global__ void kernel2(float *dranks, int *dlinks, float *dlink_weights, int nD
 
 extern "C" {
     
-    void prCUDA1(float* ranks, int *links, int *link_counts, float *link_weights, int nDocs, int nLinks, int start, int end, int GPUN) {
+    void prCUDA1(float* ranks, int *links, int *link_counts, float *link_weights, long nDocs, long nLinks, long start, long end, long GPUN) {
 	float *dranks, *dlink_weights;
 	int *dlinks, *dlink_counts;
 	if (GPUN > 0) {
@@ -84,7 +84,7 @@ extern "C" {
 	}
     }
     
-    void prCUDA2(float* ranks, int *links, float *link_weights, int nDocs, int nLinks, int start, int end, int GPUN) {
+    void prCUDA2(float* ranks, int *links, float *link_weights, long nDocs, long nLinks, long start, long end, long GPUN) {
 	float *dranks, *dlink_weights;
 	int *dlinks;
 	if (GPUN > 0) {
